worked on geocoding, broke app, think I figured out what to do next time, saving in this branch to refer to
